services/auth-service/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
python-multipart
email-validator
bcrypt==4.0.1

# curl -X 'POST' \
#   'http://localhost:8000/api/auth/register' \
#   -H 'accept: application/json' \
#   -H 'Content-Type: application/json' \
#   -d '{
#   "email": "__test__@dvfu.ru",
#   "password": "dvfu",
#   "full_name": "Evgenii Uglitckii",
#   "phone": "12345678901"
# }'